Refactored queryoptions.py
